Place the build and package assembly code into it's own file, prepare the code so it can be used in a commandlet.
