+++ SpatialGDK/Source/SpatialGDKEditor/Public/SpatialGDKEditor.h
@@ -25,6 +25,10 @@ public:
 	bool IsSchemaGeneratorRunning() { return bSchemaGeneratorRunning; }
 	bool FullScanRequired();
 
+	void BuildServerAssembly();
+	void BuildClientAssembly();
+	void BuildSimulatedPlayerAssembly();
+
 private:
 	bool bSchemaGeneratorRunning;
 	TFuture<bool> SchemaGeneratorResult;
